football_ai_pipeline/requirements.txt
numpy>=1.24,<2.0
scipy>=1.10
scikit-learn>=1.2
opencv-python>=4.7
shapely>=2.0
matplotlib>=3.7
pyyaml>=6.0
pandas>=2.0
tqdm>=4.65
# Optional — GPU-accelerated detection & tracking
# ultralytics>=8.0
# supervision>=0.18
# torch>=2.0
# umap-learn>=0.5
streamlit>=1.30
pytest>=7.0
